[#131] refactor: Coupon 도메인 이동으로 인한 수정
