+++ src/main/java/bc1/gream/domain/order/service/command/OrderCommandService.java
@@ -1,11 +1,11 @@
 package bc1.gream.domain.order.service.command;
 
 import bc1.gream.domain.buy.entity.Buy;
+import bc1.gream.domain.coupon.entity.Coupon;
+import bc1.gream.domain.coupon.helper.CouponCalculator;
 import bc1.gream.domain.order.entity.Order;
 import bc1.gream.domain.order.repository.OrderRepository;
 import bc1.gream.domain.sell.entity.Sell;
-import bc1.gream.domain.user.coupon.entity.Coupon;
-import bc1.gream.domain.user.coupon.helper.CouponCalculator;
 import bc1.gream.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -49,7 +49,7 @@ public class OrderCommandService {
             .finalPrice(finalPrice)
             .expectedPrice(sell.getPrice())
             .build();
-        
+
         return orderRepository.save(order);
     }
 }
